+++ common/src/main/kotlin/com/lambda/command/commands/ModuleCommand.kt
@@ -16,7 +16,7 @@ import com.lambda.util.Communication.info
 import com.lambda.util.Communication.joinToText
 import com.lambda.util.Communication.warn
 import com.lambda.util.StringUtils
-import com.lambda.util.primitives.extension.CommandBuilder
+import com.lambda.util.extension.CommandBuilder
 import com.lambda.util.text.ClickEvents.suggestCommand
 import com.lambda.util.text.buildText
 import com.lambda.util.text.clickEvent
