+++ common/src/main/kotlin/com/lambda/module/modules/combat/Criticals.kt
@@ -7,10 +7,10 @@ import com.lambda.interaction.rotation.Rotation
 import com.lambda.interaction.rotation.Rotation.Companion.rotationTo
 import com.lambda.module.Module
 import com.lambda.module.tag.ModuleTag
-import com.lambda.util.primitives.extension.component1
-import com.lambda.util.primitives.extension.component2
-import com.lambda.util.primitives.extension.component3
-import com.lambda.util.primitives.extension.rotation
+import com.lambda.util.extension.component1
+import com.lambda.util.extension.component2
+import com.lambda.util.extension.component3
+import com.lambda.util.extension.rotation
 import net.minecraft.network.packet.c2s.play.PlayerActionC2SPacket
 import net.minecraft.network.packet.c2s.play.PlayerInteractItemC2SPacket
 import net.minecraft.network.packet.c2s.play.PlayerMoveC2SPacket
