+++ common/src/main/kotlin/com/lambda/module/modules/movement/Speed.kt
@@ -24,7 +24,7 @@ import com.lambda.util.player.MovementUtils.newMovementInput
 import com.lambda.util.player.MovementUtils.roundedForward
 import com.lambda.util.player.MovementUtils.roundedStrafing
 import com.lambda.util.player.MovementUtils.setSpeed
-import com.lambda.util.primitives.extension.contains
+import com.lambda.util.extension.contains
 import com.lambda.util.world.WorldUtils.getFastEntities
 import net.minecraft.entity.LivingEntity
 import net.minecraft.entity.decoration.ArmorStandEntity
