+++ common/src/main/kotlin/com/lambda/task/tasks/InventoryTask.kt
@@ -9,8 +9,8 @@ import com.lambda.task.Task
 import com.lambda.util.item.ItemUtils.block
 import com.lambda.util.item.ItemUtils.defaultDisposables
 import com.lambda.util.player.SlotUtils.clickSlot
-import com.lambda.util.primitives.extension.containerSlots
-import com.lambda.util.primitives.extension.inventorySlots
+import com.lambda.util.extension.containerSlots
+import com.lambda.util.extension.inventorySlots
 import net.minecraft.item.ItemStack
 import net.minecraft.screen.ScreenHandler
 import net.minecraft.screen.slot.Slot
