+++ common/src/main/kotlin/com/lambda/util/extension/Box.kt
@@ -1,4 +1,4 @@
-package com.lambda.util.primitives.extension
+package com.lambda.util.extension
 
 import net.minecraft.util.math.Box
 import net.minecraft.util.math.Vec3d
