+++ common/src/main/kotlin/com/lambda/util/extension/Mixin.kt
@@ -1,0 +1,6 @@
+package com.lambda.util.extension
+
+import net.minecraft.client.MinecraftClient
+
+val MinecraftClient.partialTicks
+    get() = (if (paused) pausedTickDelta else tickDelta).toDouble()
