--- common/src/main/kotlin/com/lambda/util/primitives/extension/Mixin.kt
@@ -1,6 +1,0 @@
-package com.lambda.util.primitives.extension
-
-import net.minecraft.client.MinecraftClient
-
-val MinecraftClient.partialTicks
-    get() = (if (paused) pausedTickDelta else tickDelta).toDouble()
